refs CRASM-3382: Move main component tests to centralized location
- Move 8 component test files from src/components/__tests__/ to src/tests/components/
- Rename .spec.tsx files to .test.tsx for consistency
- Update import paths to reflect new file locations
- Remove old snapshot files and empty __tests__ directory
