+++ frontend/src/tests/components/domainDetails.test.tsx
