+++ frontend/src/tests/components/domainsTable.test.tsx
