+++ frontend/src/tests/components/layout.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'test-utils/test-utils';
 import { afterAll, describe, expect, it, vi } from 'vitest';
-import { Layout } from '../Layout';
+import { Layout } from '../../components/Layout';
 import { StaticsContext, StaticsContextType } from 'context/StaticsContext';
 import {
   ContextType,
