+++ frontend/src/tests/components/routeGuard.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as router from 'react-router-dom';
-import { RouteGuard } from '../Routes/RouteGuard';
+import { RouteGuard } from '../../components/Routes/RouteGuard';
 import { render, testUser } from 'test-utils';
 import { afterAll, afterEach, beforeEach, expect, it, vi } from 'vitest';
