+++ frontend/src/tests/components/updateUserStateForm.test.tsx
@@ -7,7 +7,7 @@ import {
 } from '../../test-utils/test-utils';
 import { testUser } from '../../test-utils/user';
 import { describe, expect, it, vi, beforeEach } from 'vitest';
-import { UpdateStateForm } from '../UpdateUserStateForm/UpdateUserStateForm';
+import { UpdateStateForm } from '../../components/UpdateUserStateForm/UpdateUserStateForm';
 
 vi.mock('@/constants/constants', () => ({
   STATE_OPTIONS: ['California', 'Texas', 'New York', 'Florida']
