+++ frontend/src/tests/components/vulnerabilitiesTable.test.tsx
